Bump regex version
In the transition from regex 0.1 to 0.2 the short form for
the Unicode currency class became unsupported. This patch
switches claude to the longer form and switches the parsing
code to use `captures` instead of `captures_iter`, because
there was a comment asking for the change. We also switch
to using an anchored match instead of explicitly checking to
see if the regex matched the whole input.